feat(T-205~T-207): observer hooks + Stage 0 + cross-session linkage
T-205 5 adapter hook scripts + engram observer install
- adapters/<client>/hooks/engram_observe_post_tool_use.sh × 5
  (claude-code / codex / cursor / gemini-cli / opencode)
- engram/observer/translators.py — Claude Code / Codex hook payload
  → engram event mapper (tool_name + tool_input.file_path/files →
  event=tool_use + files; truncates large tool_response to char count)
- engram/observer/install.py — InstallTarget registry + plan + apply,
  write/paste actions; _adapters_root anchors on adapters/claude-code/
  hooks/ to avoid colliding with cli/engram/adapters/ Python module
- engram/observer/install_cli.py — `engram observer` group + `install`
  --target=<client> --list --dry-run --format=json (claude-code writes
  ~/.claude/settings.json; others paste); idempotent re-install
- engram observe --from=<client> stdin payload translation
- 31 unit tests

T-206 Relevance Gate Stage 0
- engram/relevance/gate.py: SessionContinuation dataclass +
  select_session_continuations pure fn + Stage 0 in run_relevance_gate
- task_hash match → same-hash sessions sorted by ended_at desc, capped
  at max_count=3 + budget_fraction=0.25; task_hash=None is no-op
- engram/observer/loader.py — disk Session asset → SessionContinuation
  (skips sessions without task_hash since they cannot match)
- 15 unit tests

T-207 Cross-session task linkage
- engram/observer/linkage.py: find_predecessor (same hash, ended_at <
  self.started_at, excludes self), set_next_session (fcntl.flock
  LOCK_EX + write_atomic), link_session_to_predecessor (writes both
  ends, back-ref failure non-fatal)
- Tier 1 compact_to_session_asset auto-links when task_hash supplied,
  with contextlib.suppress(Exception) — linkage failure never kills
  Tier 1 write
- 15 unit tests, including E2E across 2 sessions verifying prev/next
  bidirectional pointers

Integrity:
- 959 passed + 1 xfailed (+66 from 893)
- mypy strict clean on engram/observer/ + engram/relevance/
- README ## Commands updated (engram observe --from + observer install)
- HTML banner + Roadmap synced (en + zh)

Side-effects from ruff --fix --unsafe-fixes pass:
- import sorting + try/except/pass → contextlib.suppress conversions
  in mcp/install.py, commands/context.py and other touched files;
  no behaviour change.

+++ cli/engram/consistency/phase2_semantic.py
@@ -187,7 +187,7 @@ def detect_phase2(store_root: Path) -> list[ConflictReport]:
             sim = _jaccard(sh_a, sh_b)
             if sim < _FUZZY_MERGE_THRESHOLD:
                 continue
-            pct = int(round(sim * 100))
+            pct = round(sim * 100)
             reports.append(
                 ConflictReport(
                     conflict_class=ConflictClass.FACTUAL,
